requirements.txt
regex
requests
GitPython
PyYAML